+++ test/src/test_little_endian.cpp
@@ -3,8 +3,12 @@
 //
 // Distributed under the "BSD License". See the accompanying LICENSE.rst file.
 
-#include <endian/is_big_endian.hpp>
 #include <endian/little_endian.hpp>
+
+#include <cstdint>
+
+#include <endian/is_big_endian.hpp>
+
 #include <gtest/gtest.h>
 
 
@@ -31,13 +35,13 @@ TEST(test_little_endian, convert)
     // Test 16-bit integer
     {
         uint8_t data[2];
-        uint16_t input = 0x2211U;
+        uint16_t input = 0x1122U;
 
         // If the host is little endian, the put function
         // should change the byte order (no change for big endian)
         endian::little_endian::put16(input, data);
-        EXPECT_EQ(0x22U, data[1]);
-        EXPECT_EQ(0x11U, data[0]);
+        EXPECT_EQ(0x22U, data[0]);
+        EXPECT_EQ(0x11U, data[1]);
 
         // Get should swap the value back (no change for big endian)
         uint16_t out = endian::little_endian::get16(data);
@@ -47,15 +51,15 @@ TEST(test_little_endian, convert)
     // Test 32-bit integer
     {
         uint8_t data[4];
-        uint32_t input = 0x44332211U;
+        uint32_t input = 0x11223344U;
 
         // If the host is little endian, the put function
         // should change the byte order (no change for big endian)
         endian::little_endian::put32(input, data);
-        EXPECT_EQ(0x44U, data[3]);
-        EXPECT_EQ(0x33U, data[2]);
-        EXPECT_EQ(0x22U, data[1]);
-        EXPECT_EQ(0x11U, data[0]);
+        EXPECT_EQ(0x44U, data[0]);
+        EXPECT_EQ(0x33U, data[1]);
+        EXPECT_EQ(0x22U, data[2]);
+        EXPECT_EQ(0x11U, data[3]);
 
         // Get should swap the value back (no change for big endian)
         uint32_t out = endian::little_endian::get32(data);
@@ -65,19 +69,19 @@ TEST(test_little_endian, convert)
     // Test 64-bit integer
     {
         uint8_t data[8];
-        uint64_t input = 0x8877665544332211ULL;
+        uint64_t input = 0x1122334455667788ULL;
 
         // If the host is little endian, the put function
         // should change the byte order (no change for big endian)
         endian::little_endian::put64(input, data);
-        EXPECT_EQ(0x88U, data[7]);
-        EXPECT_EQ(0x77U, data[6]);
-        EXPECT_EQ(0x66U, data[5]);
-        EXPECT_EQ(0x55U, data[4]);
-        EXPECT_EQ(0x44U, data[3]);
-        EXPECT_EQ(0x33U, data[2]);
-        EXPECT_EQ(0x22U, data[1]);
-        EXPECT_EQ(0x11U, data[0]);
+        EXPECT_EQ(0x88U, data[0]);
+        EXPECT_EQ(0x77U, data[1]);
+        EXPECT_EQ(0x66U, data[2]);
+        EXPECT_EQ(0x55U, data[3]);
+        EXPECT_EQ(0x44U, data[4]);
+        EXPECT_EQ(0x33U, data[5]);
+        EXPECT_EQ(0x22U, data[6]);
+        EXPECT_EQ(0x11U, data[7]);
 
         // Get should swap the value back (no change for big endian)
         uint64_t out = endian::little_endian::get64(data);