+++ test/src/test_stream.cpp
@@ -3,117 +3,222 @@
 //
 // Distributed under the "BSD License". See the accompanying LICENSE.rst file.
 
+#include <endian/stream.hpp>
 
-#include <cstdlib>
+#include <cstdint>
 #include <vector>
-#include <string>
-#include <algorithm>
 
-#include "../helper_functions.hpp"
+#include <endian/stream_reader.hpp>
+#include <endian/stream_writer.hpp>
 #include <endian/big_endian.hpp>
 #include <endian/little_endian.hpp>
 
-#include <endian/stream_writer.hpp>
-#include <endian/stream_reader.hpp>
 #include <gtest/gtest.h>
 
-template<class EndianType>
-static void run_test_create_writer()
+
+TEST(test_stream, basic_api)
 {
-    const uint32_t size = 1024;
+    {
+        SCOPED_TRACE(testing::Message() << "size 1");
+        uint32_t size = 1U;
+        endian::stream stream(size);
+
+        // check initial state
+        EXPECT_EQ(size, stream.size());
+        EXPECT_EQ(0U, stream.position());
+        EXPECT_EQ(size, stream.remaining());
+
+        // check state after seek
+        stream.seek(1);
+        EXPECT_EQ(size, stream.size());
+        EXPECT_EQ(1U, stream.position());
+        EXPECT_EQ(0U, stream.remaining());
+    }
+
+    {
+        SCOPED_TRACE(testing::Message() << "size 10000");
+        uint32_t size = 10000U;
+        endian::stream stream(size);
+
+        // check initial state
+        EXPECT_EQ(size, stream.size());
+        EXPECT_EQ(0U, stream.position());
+        EXPECT_EQ(size, stream.remaining());
+
+        // check state after seek
+        stream.seek(size / 2);
+        EXPECT_EQ(size, stream.size());
+        EXPECT_EQ(size / 2, stream.position());
+        EXPECT_EQ(size / 2, stream.remaining());
+
+        // that consecutive seeks doesn't alter state.
+        stream.seek(size / 2);
+        stream.seek(size / 2);
+        stream.seek(size / 2);
+        EXPECT_EQ(size, stream.size());
+        EXPECT_EQ(size / 2, stream.position());
+        EXPECT_EQ(size / 2, stream.remaining());
+
+        // seek to end
+        stream.seek(size);
+        EXPECT_EQ(size, stream.position());
+        EXPECT_EQ(0U, stream.remaining());
+    }
+}
+
+// ********************** //
+//  Reader writer tests:  //
+// ********************** //
+
+template<class ValueType, class EndianType>
+void write_and_read_value_type_test()
+{
+    const uint32_t elements = 1024;
+    const uint32_t size = 1024 * sizeof(ValueType);
+    std::vector<uint8_t> buffer;
+    buffer.resize(size);
+    endian::stream_reader<EndianType> stream_reader(buffer.data(), size);
+    endian::stream_writer<EndianType> stream_writer(buffer.data(), size);
+
+    ValueType lowest_value = std::numeric_limits<ValueType>::min();
+    ValueType highest_value = std::numeric_limits<ValueType>::max();
+
+    for (uint32_t i = 0; i < elements; i++)
+    {
+        stream_writer.write(highest_value);
+    }
+
+    stream_reader.seek(0);
+    for (uint32_t i = 0; i < elements; i++)
+    {
+        ValueType value = 0;
+        stream_reader.read(value);
+        EXPECT_EQ(highest_value, value);
+    }
+
+    stream_writer.seek(0);
+    for (uint32_t i = 0; i < elements; i++)
+    {
+        stream_writer.write(lowest_value);
+    }
+
+    stream_reader.seek(0);
+    for (uint32_t i = 0; i < elements; i++)
+    {
+        ValueType value = 0;
+        stream_reader.read(value);
+        EXPECT_EQ(lowest_value, value);
+    }
+}
+
+template<class ValueType, class EndianType>
+void write_and_read_random_value_type_test()
+{
+    const uint32_t elements = 1024;
+    const uint32_t size = 1024 * sizeof(ValueType);
     std::vector<uint8_t> buffer;
     buffer.resize(size);
 
     endian::stream_writer<EndianType> stream_writer(buffer.data(), size);
-
-    EXPECT_EQ(size, stream_writer.size());
-    EXPECT_EQ(0u, stream_writer.position());
-}
-
-template<class EndianType>
-static void run_test_create_reader()
-{
-    const uint32_t size = 1024;
-    std::vector<uint8_t> buffer;
-    buffer.resize(size);
-
     endian::stream_reader<EndianType> stream_reader(buffer.data(), size);
 
-    EXPECT_EQ(size, stream_reader.size());
-    EXPECT_EQ(0u, stream_reader.position());
-}
+    ValueType highest_value = std::numeric_limits<ValueType>::max();
 
-template<class EndianType>
-static void run_test_simple_read_write()
-{
-    const uint32_t elements = 10; /// no. of elements
-    const uint32_t size = elements * sizeof(uint32_t);
+    std::vector<ValueType> values;
+    values.resize(elements);
 
-    std::vector<uint8_t> buffer;
-    buffer.resize(size);
-
-    endian::stream_writer<EndianType> writer(buffer.data(), buffer.size());
-
-    endian::stream_reader<EndianType> reader(buffer.data(), buffer.size());
-
-    EXPECT_EQ(size, writer.size());
-    EXPECT_EQ(0U, writer.position());
-
-    EXPECT_EQ(size, reader.size());
-    EXPECT_EQ(0U, reader.position());
+    uint32_t seed = static_cast<uint32_t>(time(0));
+    SCOPED_TRACE(testing::Message() << "seed: " << seed);
+    srand(seed);
 
     for (uint32_t i = 0; i < elements; i++)
     {
-        writer.write(i);
+        values[i] = rand() % (highest_value);
+        stream_writer.write(values[i]);
     }
-
-    EXPECT_EQ(size, writer.size());
-    EXPECT_EQ(0U, writer.position());
-
-    uint32_t last_value = 0;
-    for(uint32_t i = 0; i < elements; i++)
+    for (uint32_t i = 0; i < elements; i++)
     {
-        reader.read(last_value);
-        EXPECT_EQ(i, last_value);
+        ValueType value;
+        stream_reader.read(value);
+        EXPECT_EQ(values[i], value);
     }
 }
 
-/// Write-read test
-template<class ValueType, class EndianType>
-static void run_write_read_test()
-{
-    write_read_test<ValueType, EndianType>();
-}
-
-/// Pseudorandom write-read tests
-template<class ValueType, class EndianType>
-static void run_pseudorandom_read_write_test()
-{
-    random_write_read_test<ValueType, EndianType>(true);
-}
-
-/// Random write read tests
-template<class ValueType, class EndianType>
-static void run_random_read_write_test()
-{
-    random_write_read_test<ValueType, EndianType>(false);
-}
-
-/// Various read writes
 template<class EndianType>
-static void run_pseudorandom_various_read_writes_test()
+void write_and_read_variadic_types_test()
 {
-    various_write_read_test<EndianType>(true);
+    const uint32_t elements = 1024;
+    const uint32_t size = 1024 * sizeof(uint64_t);
+    std::vector<uint8_t> buffer;
+    buffer.resize(size);
+
+    endian::stream_writer<EndianType> writer(buffer.data(), size);
+
+    std::vector<uint64_t> values;
+    values.resize(elements);
+
+    uint32_t seed = static_cast<uint32_t>(time(0));
+    SCOPED_TRACE(testing::Message() << "seed: " << seed);
+    srand(seed);
+
+    for (uint32_t i = 0; i < elements; i++)
+    {
+        switch (i % 4)
+        {
+            case 0:
+                values[i] = rand() % std::numeric_limits<uint8_t>::max();
+                writer.write((uint8_t)values[i]);
+                break;
+            case 1:
+                values[i] = rand() % std::numeric_limits<uint16_t>::max();
+                writer.write((uint16_t)values[i]);
+                break;
+            case 2:
+                values[i] = rand() % std::numeric_limits<uint32_t>::max();
+                writer.write((uint32_t)values[i]);
+                break;
+            case 3:
+                values[i] = rand() % std::numeric_limits<uint64_t>::max();
+                writer.write((uint64_t)values[i]);
+                break;
+        }
+    }
+
+    uint8_t last_u8 = 0;
+    uint16_t last_u16 = 0;
+    uint32_t last_u32 = 0;
+    uint64_t last_u64 = 0;
+
+    // create reader
+    endian::stream_reader<EndianType> reader(buffer.data(), size);
+
+    // Read values in FIFO order
+    for (uint32_t i = 0; i < elements; i++)
+    {
+        switch (i % 4)
+        {
+            case 0:
+                reader.read(last_u8);
+                EXPECT_EQ(values[i], last_u8);
+                break;
+            case 1:
+                reader.read(last_u16);
+                EXPECT_EQ(values[i], last_u16);
+                break;
+            case 2:
+                reader.read(last_u32);
+                EXPECT_EQ(values[i], last_u32);
+                break;
+            case 3:
+                reader.read(last_u64);
+                EXPECT_EQ(values[i], last_u64);
+                break;
+        }
+    }
 }
 
 template<class EndianType>
-static void run_random_various_read_writes_test()
-{
-    various_write_read_test<EndianType>(false);
-}
-
-template<class EndianType>
-static void run_read_write_string_test()
+static void write_and_read_string_test()
 {
     const uint32_t size = 1024;
     std::vector<uint8_t> buffer;
@@ -134,34 +239,38 @@ static void run_read_write_string_test()
     writer.write((uint16_t)third.size());
     writer.write((uint8_t*)third.c_str(), third.size());
 
-    // Temp variables
-    std::string current;
-    uint16_t len = 0;
-
     // Create reader
     endian::stream_reader<EndianType> reader(buffer.data(), size);
 
     // Read the strings together with their lengths
-    reader.read(len);
-    EXPECT_EQ(first.size(), len);
-    // Resize the current string to accommodate 'len' bytes
-    current.resize(len);
-    reader.read((uint8_t*)current.c_str(), len);
-    EXPECT_EQ(first, current);
-
-    reader.read(len);
-    EXPECT_EQ(second.size(), len);
-    current.resize(len);
-    reader.read((uint8_t*)current.c_str(), len);
-    EXPECT_EQ(second, current);
-
-    reader.read(len);
-    EXPECT_EQ(third.size(), len);
-    current.resize(len);
-    reader.read((uint8_t*)current.c_str(), len);
-    EXPECT_EQ(third, current);
+    {
+        std::string string;
+        uint16_t length;
+        reader.read(length);
+        EXPECT_EQ(first.size(), length);
+        string.resize(length);
+        reader.read((uint8_t*)string.c_str(), length);
+        EXPECT_EQ(first, string);
+    }
+    {
+        std::string string;
+        uint16_t length;
+        reader.read(length);
+        EXPECT_EQ(second.size(), length);
+        string.resize(length);
+        reader.read((uint8_t*)string.c_str(), length);
+        EXPECT_EQ(second, string);
+    }
+    {
+        std::string string;
+        uint16_t length;
+        reader.read(length);
+        EXPECT_EQ(third.size(), length);
+        string.resize(length);
+        reader.read((uint8_t*)string.c_str(), length);
+        EXPECT_EQ(third, string);
+    }
 }
-
 
 template<class EndianType>
 static void run_read_write_vector_test()
@@ -171,93 +280,88 @@ static void run_read_write_vector_test()
     buffer.resize(size);
     endian::stream_writer<EndianType> writer(buffer.data(), size);
 
-    std::vector<uint8_t> first(100, 'a');
+    using first_type = uint8_t;
+    std::vector<first_type> first(100, 'a');
     using second_type = uint32_t;
     std::vector<second_type> second(200, 1234);
 
     // Write the vectors together with their lengths
     // The length is written as 16-bit integers
     writer.write((uint16_t)first.size());
-    writer.write(first.data(), first.size());
+    writer.write((uint8_t*)first.data(), first.size() * sizeof(first_type));
+
     // The size here refers to the number of integers
     // stored in the second vector
     writer.write((uint16_t)second.size());
     writer.write((uint8_t*)second.data(), second.size() * sizeof(second_type));
 
     // Temp variables
-    std::vector<uint8_t> first_out;
+    std::vector<first_type> first_out;
     std::vector<second_type> second_out;
-    uint16_t len = 0;
 
     // Create reader
     endian::stream_reader<EndianType> reader(buffer.data(), size);
 
-    // Read the vector length
-    reader.read(len);
-    EXPECT_EQ(first.size(), len);
-    // Resize the output vector to accommodate 'len' bytes
-    first_out.resize(len);
-    reader.read(first_out.data(), len);
-    EXPECT_EQ(first, first_out);
+    {
+        uint16_t length;
+        reader.read(length);
+        EXPECT_EQ(first.size(), length);
 
-    // Read the vector length
-    reader.read(len);
-    EXPECT_EQ(second.size(), len);
-    // Resize the output vector to accommodate 'len' bytes
-    second_out.resize(len);
-    reader.read((uint8_t*)second_out.data(), len * sizeof(second_type));
-    EXPECT_EQ(second, second_out);
+        std::vector<first_type> vector;
+        vector.resize(length);
+        reader.read((uint8_t*)vector.data(), length * sizeof(first_type));
+        EXPECT_EQ(first, vector);
+    }
+    {
+        uint16_t length;
+        reader.read(length);
+        EXPECT_EQ(second.size(), length);
+
+        std::vector<second_type> vector;
+        vector.resize(length);
+        reader.read((uint8_t*)vector.data(), length * sizeof(second_type));
+        EXPECT_EQ(second, vector);
+    }
 }
 
 template<class EndianType>
-static void test_basic_api()
+static void test_reader_and_writer_api()
 {
-    run_test_create_writer<EndianType>();
-
-    run_test_create_reader<EndianType>();
-
-    SCOPED_TRACE(testing::Message() << "uint = uint8_t");
     {
-        run_write_read_test<uint8_t, EndianType>();
-        run_pseudorandom_read_write_test<uint8_t, EndianType>();
-        run_random_read_write_test<uint8_t, EndianType>();
+        SCOPED_TRACE("uint8_t");
+        write_and_read_value_type_test<uint8_t, EndianType>();
+        write_and_read_random_value_type_test<uint8_t, EndianType>();
+    }
+    {
+        SCOPED_TRACE("uint16_t");
+        write_and_read_value_type_test<uint16_t, EndianType>();
+        write_and_read_random_value_type_test<uint16_t, EndianType>();
+    }
+    {
+        SCOPED_TRACE("uint32_t");
+        write_and_read_value_type_test<uint32_t, EndianType>();
+        write_and_read_random_value_type_test<uint32_t, EndianType>();
+    }
+    {
+        SCOPED_TRACE("uint64_t");
+        write_and_read_value_type_test<uint64_t, EndianType>();
+        write_and_read_random_value_type_test<uint64_t, EndianType>();
     }
 
-    SCOPED_TRACE(testing::Message() << "uint = uint16_t");
-    {
-        run_write_read_test<uint16_t, EndianType>();
-        run_pseudorandom_read_write_test<uint16_t, EndianType>();
-        run_random_read_write_test<uint16_t, EndianType>();
-    }
-
-    SCOPED_TRACE(testing::Message() << "uint = uint32_t");
-    {
-        run_write_read_test<uint32_t, EndianType>();
-        run_pseudorandom_read_write_test<uint32_t, EndianType>();
-        run_random_read_write_test<uint32_t, EndianType>();
-    }
-
-    SCOPED_TRACE(testing::Message() << "uint = uint64_t");
-    {
-        run_write_read_test<uint64_t, EndianType>();
-        run_pseudorandom_read_write_test<uint64_t, EndianType>();
-        run_random_read_write_test<uint64_t, EndianType>();
-    }
-
-    run_pseudorandom_various_read_writes_test<EndianType>();
-    run_random_various_read_writes_test<EndianType>();
-
-    run_read_write_string_test<EndianType>();
+    write_and_read_variadic_types_test<EndianType>();
+    write_and_read_string_test<EndianType>();
     run_read_write_vector_test<EndianType>();
 }
 
-
-TEST(test_stream, big_endian)
+TEST(test_stream, test_reader_and_writer)
 {
-    test_basic_api<endian::big_endian>();
-}
+    {
+        SCOPED_TRACE("big endian");
+        test_reader_and_writer_api<endian::big_endian>();
+    }
 
-TEST(test_stream, little_endian)
-{
-    test_basic_api<endian::little_endian>();
+    {
+        SCOPED_TRACE("little endian");
+        test_reader_and_writer_api<endian::little_endian>();
+    }
 }